requirements.txt
aiofile==3.3.3
aiohttp<3.7.0
async-timeout==3.0.1
attrs==20.2.0
caio==0.6.3
cchardet==2.1.7
cffi==1.14.3
chardet==3.0.4
chronous==2.0.0b3
colorama==0.4.4
colorlog==4.6.2
discord.py==1.6.0
multidict==4.7.6
pycparser==2.20
pydest==0.4.0
PyNaCl==1.3.0
six==1.15.0
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
yarl==1.5.1
